% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{assign_taxonomy}
\alias{assign_taxonomy}
\title{Assign full taxonomy to species}
\usage{
assign_taxonomy(kraken.tax, species.tab)
}
\arguments{
\item{kraken.tax}{This is a kraken2 report generated by running kraken2 using the arguments --use-mpa-style and --report-zero-counts with the desired kraken2 database on a test sample. The first column of this report should be renamed "Taxonomy".
An example of the kraken2 output can be seen with data("krakenmet").}

\item{species.tab}{The table to which taxonomy whould be assigned. Samples are rows and defined in the column "sample" and species are columns.}
}
\value{
A long formatted table with each species taxonomy reported in a column for each taxonomic level.
}
\description{
Using a kraken2 report with metaphlan formatted output, assign to the normalized table resulting from voom_snm_normalization.
This is also compatible with the unnormalized counts and rarefied prevalence outputs.
}
