% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_abundance_relative_to_human.R
\name{calculate_abundance_relative_to_human}
\alias{calculate_abundance_relative_to_human}
\title{Calculate abundance relative to amount of human of unnormalized species counts}
\usage{
calculate_abundance_relative_to_human(counts)
}
\arguments{
\item{counts}{Table with normalized or unnormalized species counts. Samples are expected to be in a column labelled "sample", Human counts are expected to be in a column labelled "Homo.sapiens"}
}
\value{
A table where rows are samples, columns are species, and cells are the relative abundances.
}
\description{
Given a table where rows are samples and columns are species, calculate the relative abundances for each sample.
}
