% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_relative_abundance.R
\name{calculate_relative_abundance}
\alias{calculate_relative_abundance}
\title{Calculate relative abundances of normalized species counts}
\usage{
calculate_relative_abundance(counts)
}
\arguments{
\item{counts}{Table with normalized or unnormalized species counts. Samples are expected to be in a column labelled "sample".}
}
\value{
A table where rows are samples, columns are species, and cells are the relative abundances.
}
\description{
Given a table where rows are samples and columns are species, calculate the relative abundances for each sample.
}
