% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_human_percentages.R
\name{check_human_percentages}
\alias{check_human_percentages}
\title{Remove samples with low human abundances}
\usage{
check_human_percentages(counts, req.percent)
}
\arguments{
\item{counts}{A table where rows are samples as defined in the column "sample", columns are species, and values are counts.}

\item{req.precent}{The percentage of a sample that must be human to keep the sample.}
}
\value{
A data frame containing only the samples that contain at least the required percentage of human data.
}
\description{
The exotic pipeline was built with the intention of processing human tumor samples. Samples with low human abundances are therefore suspiscious and should be removed.
}
