% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_rarefied_prevalence.R
\name{counts_to_rarefied_prevalence}
\alias{counts_to_rarefied_prevalence}
\title{Calculate rarefied prevalence of count data}
\usage{
counts_to_rarefied_prevalence(counts, seed)
}
\arguments{
\item{counts}{A data table where rows are samples as defined in the column "sample", columns are species, and values are unnormalized counts.}

\item{seed}{Seed used for rarefying samples.}
}
\value{
A table containing the rarefied prevalences for the counts table provided.
}
\description{
Normalizing data with VOOM-SNM removes any zero counts to maintin compatibility with a log transformation,
so for analyses requiring prevalence data it is better to use the rarefied prevalence of the unnormalized count data.
This function rarefies to match the depth of the sample with the minimum counts in the provided table.
}
