% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_voom_snm_expr.R
\name{format_voom_snm_expr}
\alias{format_voom_snm_expr}
\title{Formatting TCGA gene expression data for normalization}
\usage{
format_voom_snm_expr(tcga.exp, tcga.meta)
}
\arguments{
\item{tcga.exp}{A table of expression values where rows are genes with gene names stored as row names and columns are samples.
It is expected that the TCGA sample names will need to be reformatted by the function from R compatible column names back the the file ids TCGA provides.}

\item{tcga.meta}{The results of the function grab_TCGA_metadata.}
}
\value{
A list of the required objects for normalization with the counts listed first and the meta data listed second.
}
\description{
This function is meant to assist with formatting TCGA data processed with other exotic functions for voom-snm normalization.
}
