% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_voom_snm_microbe.R
\name{format_voom_snm_microbe}
\alias{format_voom_snm_microbe}
\title{Formatting TCGA microbe count data for normalization}
\usage{
format_voom_snm_microbe(tcga.counts, tcga.meta)
}
\arguments{
\item{tcga.counts}{A table of microbe counts where rows are samples with sample names stored in the column "sample" and columns are species.}

\item{tcga.meta}{The results of the function grab_TCGA_metadata.}
}
\value{
A list of the required objects for normalization with the counts listed first and the meta data listed second.
}
\description{
This function is meant to assist with formatting TCGA data processed with other exotic functions for voom-snm normalization.
}
