% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_TCGA_manifest.R
\name{generate_TCGA_manifest}
\alias{generate_TCGA_manifest}
\title{Generate genomic data commons manifest}
\usage{
generate_TCGA_manifest(cname)
}
\arguments{
\item{cname}{The name of the TCGA cancer project to obtain samples ids for.}
}
\value{
A data table of the file ids for both aliged reads and gene expression results for the given cancer. These can then be passed to tools like the gdc client to download data files.
}
\description{
This function interacts with the Genomic Data Commons API to obtain information on file ids for a given TCGA project.
These are filtered to RNA-seq samples of the primary tumor.
}
