% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contaminants.R
\name{get_contaminants}
\alias{get_contaminants}
\title{Determine statistical contaminants}
\usage{
get_contaminants(metadf, countdf, seed)
}
\arguments{
\item{metadf}{Data frame containing the RNA concentration and batch information for the samples. These are expected in the columns
"rna.conc" and "batch" respectively. Sample order must match the sample order of countdf.}

\item{countdf}{Data frame where rows are samples with the sample names stored in the column "sample" and columns are species.}

\item{seed}{Seed value for sampling in the analysis.}
}
\value{
A data frame indicating whether each species is a likely contaminant, with an additional column indicating the genus
each species belongs to.
}
\description{
Determine statistical contaminants using RNA concentrations and sample batches.
}
