% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_TCGA_clinical.R
\name{grab_TCGA_clinical}
\alias{grab_TCGA_clinical}
\title{Acquire TCGA clinical data}
\usage{
grab_TCGA_clinical(filemanifest)
}
\arguments{
\item{filemanifest}{Either the result of running generate_TCGA_manifest once or binding the results of several cancers together.}
}
\value{
A data frame of the clinical data available from TCGA for the samples in the manifest.
}
\description{
This function interacts with Genomic Data Commons to acquire TCGA clinical data for the list of samples provided in a file manifest
as generated by the function generate_TCGA_manifest.
}
