% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_TCGA_metadata.R
\name{grab_TCGA_metadata}
\alias{grab_TCGA_metadata}
\title{Acquire TCGA metadata}
\usage{
grab_TCGA_metadata(tcga.file.ids, tcga.expression.ids)
}
\arguments{
\item{tcga.file.ids}{Equivalent to the file_id.BAM column in a manifest generated by generate_TCGA_manifest.}

\item{tcga.expression.ids}{Equivalent to the file_id.expression column in a manifest generated by generate_TCGA_manifest.
The order of these IDs should match tcga.file.ids.}
}
\value{
A data frame containing a variety of meta data about the designated samples.
}
\description{
This function interacts with Genomic Data Commons to pull the metadata used for normalization.
}
