% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_batches.R
\name{resolve_batches}
\alias{resolve_batches}
\title{Prepare data for get_contaminants}
\usage{
resolve_batches(
  metadf,
  countdf,
  conc.col.name,
  sample.col.name,
  batch.col.name
)
}
\arguments{
\item{metadf}{The data frame containing batch and concentration information for the samples}

\item{countdf}{The data from containing microbe counts information, where samples are rows with sample names saved in the column "sample" and columns are microbes.}

\item{conc.col.name}{The name of the column containing the concentration information in metadf.}

\item{sample.col.name}{The name of the column containing sample information in metadf.}

\item{batch.col.name}{The name of the column containing batch information in metadf.}
}
\value{
A list of the count and meta data formatted to be taken directly by get_contaminants. Batches with only one sample are removed.
}
\description{
This function is meant to handle the data manipulation required to prepare sample and meta data for the get_contaminants function.
}
