% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_genome_filter.R
\name{transcript_genome_filter}
\alias{transcript_genome_filter}
\title{Filter species falsely identified in the fragmented CHM1 transcript & genome and mouse GRCm39 transcript & genome}
\usage{
transcript_genome_filter(counts, filters)
}
\arguments{
\item{counts}{A table where rows are samples as defined in the column "sample", columns are species, and values are counts.}

\item{filters}{A list of the filters desired - options include: humanRNA, humanWGS, mouseRNA, mouseWGS}
}
\value{
A data frame containing only the species not included in the desired filters.
}
\description{
Filter out potential human and mouse contaminants.
}
