% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom_snm_normalization.R
\name{voom_snm_normalization}
\alias{voom_snm_normalization}
\title{Normalized decontaminanted counts}
\usage{
voom_snm_normalization(
  qcMetadata,
  qcData,
  biovars,
  adjvars,
  return.format = "logcpm",
  seed
)
}
\arguments{
\item{qcMetadata}{Data frame containing all the biological and technical variables. This must have the same sample order as the count data.}

\item{qcData}{Data frame containing the count data to be normalized. Samples are rows, with sample names stored as row names.}

\item{biovars}{A vector of the names of the columns in qcMetadata containing biological variables.}

\item{adjvars}{A vector of the names of the columns in qcMetadata containing technical variables.}

\item{return.format}{Should the data be return in logged counts per million or counts per million (cpm)?}

\item{seed}{The seed used to determine randomization for the normalization process.}
}
\value{
A data from containing the normalized logged cpm or normalized cpm of the sample values.
}
\description{
Normalizes count data using voom-snm. This removes variation from specified technical variables while preserving variation from specified biological variables.
For projects only utilizing TCGA data, the functions format_voom_snm_microbe and format_voom_snm_expr automatically format inputs for this function.
}
